Prevent false positives during fuzzing by waiting for server's CRYPTO frames before completing handshake or acknowledging server's CRYPTO frame.
After fuzzing for a particular mutation for some time, we try to establish connection with the server to check if the server supports a new connection. If the connection is not established, the fuzzer indicates discovery of a vulnerability. In some cases, there is a delay when the server sends CRYPTO frames to the client and if the client tries to complete the handshake or acknowledge server's CRYPTO frames, it throws an error. with this patch, when the fuzzer tries to establish connection, it waits up to 2 seconds to receive the server's CRYPTO frames.